refactor : 토너먼트 패배팀 계산 로직 분리
