+++ src/1_Page/Championship/ui/DashBoard/ui/TeamAndMatchContainer/util/convertToTournamentFormat.ts
@@ -2,7 +2,7 @@ export const convertToTournamentFormat = (
   matches: ChampionshipMatchList[],
   teamList: ChampionshipTeamInfo[],
   championship_type: number
-): [convertedData: TournamentData[], eliminatedTeams: number[]] => {
+): TournamentData[] => {
   // championship_type 기반 bracket 크기 결정 (없으면 teamList.length 사용)
   const roundSizeMap: Record<number, number> = {
     1: 16,
@@ -196,21 +196,5 @@ export const convertToTournamentFormat = (
     });
   }
 
-  const eliminatedTeams: number[] = [];
-
-  matches.forEach((match) => {
-    if (
-      match.championship_match_first.match_team_stats_our_score >
-      match.championship_match_second.match_team_stats_our_score
-    ) {
-      eliminatedTeams.push(match.championship_match_second.team_list_idx);
-    } else if (
-      match.championship_match_first.match_team_stats_our_score <
-      match.championship_match_second.match_team_stats_our_score
-    ) {
-      eliminatedTeams.push(match.championship_match_first.team_list_idx);
-    }
-  });
-
-  return [result, eliminatedTeams];
+  return result;
 };
